Add app container
