+++ app/client/main.js
@@ -3,10 +3,10 @@ import React from 'react';
 import { Meteor } from 'meteor/meteor';
 import { render } from 'react-dom';
 import { Sites, OpenshiftEnvs, Themes } from '../imports/api/collections';
-import App from '../imports/ui/App';
+import AppContainer from '../imports/ui/containers/AppContainer';
  
 Meteor.startup(() => {
-  render(<App />, document.getElementById('render-target'));
+  render(<AppContainer />, document.getElementById('render-target'));
   Tequila.start();
 });
